Refactor Buttons’s file & folder structure
