+++ src/components/common/Buttons/ButtonSet/styles.scss
@@ -1,6 +1,3 @@
-// Global styles
-@import "../../../styles/base/settings";
-
 // Button set
 .button-set {
   $element-spacing: 0.25rem;
@@ -15,14 +12,5 @@
 
   > :not(:first-child) {
     margin-left: $element-spacing;
-  }
-}
-
-// Button list
-.button-list {
-  .text {
-    @media (max-width: $breakpoint-sm - 1) {
-      display: none;
-    }
   }
 }
